+++ backend/beets_flask/server/routes/monitor.py
@@ -1,8 +1,6 @@
 from quart import Blueprint
 from rq.job import Job
-from rq.registry import clean_registries
-from rq.worker import Worker
-from rq.worker_registration import clean_worker_registry
+from rq.worker import BaseWorker, Worker
 
 from beets_flask.redis import queues, redis_conn
 
@@ -48,7 +46,7 @@ async def get_worker_status():
         dict: A dictionary containing the status of each worker.
 
     """
-    workers: list[Worker] = Worker.all(connection=redis_conn)
+    workers: list[BaseWorker] = Worker.all(connection=redis_conn)
 
     ret_dict = {}
     for w in workers: